+++ src/main/java/cn/haitaoss/tinyspringmvc/framework/HandlerAdapter.java
@@ -1,0 +1,10 @@
+package cn.haitaoss.tinyspringmvc.framework;
+
+/**
+ * @author haitao.chen
+ * email haitaoss@aliyun.com
+ * date 2021-04-25 11:06
+ *
+ */
+public interface HandlerAdapter {
+}