+++ src/main/java/cn/haitaoss/tinyspringmvc/framework/handlerMapping/AnnotationHandlerMapping.java
@@ -14,33 +14,10 @@ import java.util.Map;
  * date 2021-04-24 18:47
  *
  */
-public class AnnotationHandleMapping extends AbstractHandlerMapping {
+public class AnnotationHandlerMapping extends AbstractHandlerMapping {
 
-    public AnnotationHandleMapping(ApplicationContext mvcContext) {
+    public AnnotationHandlerMapping(ApplicationContext mvcContext) {
         super(mvcContext);
-    }
-
-
-    @Override
-    protected void init() {
-        // 注册map，保存url和handler的对应关系
-        registryURLAndHandler();
-        // 找到容器中定义的HandlerInterceptor
-        initHandlerInterceptors();
-    }
-
-    /**
-     * 收集容器中所有的HandlerInterceptor 的实现，保存到父类handlerInterceptors属性中
-     * @author haitao.chen
-     * email
-     * date 2021/4/25 10:11 上午
-     */
-    private void initHandlerInterceptors() {
-        try {
-            handlerInterceptors = mvcContext.getBeanFactory().getBeansForType(HandlerInterceptor.class);
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
     }
 
     /**
@@ -50,7 +27,8 @@ public class AnnotationHandleMapping extends AbstractHandlerMapping {
      * email
      * date 2021/4/24 6:54 下午
      */
-    private void registryURLAndHandler() {
+    @Override
+    protected void registryURLAndHandler() {
         // 找到容器中所有标注了 @RequestMapping 的类
         for (Map.Entry<String, BeanDefinition> entry : mvcContext.getBeanFactory().getBeanDefinitionMap().entrySet()) {
             Class clazz = entry.getValue().getBeanClass();