+++ src/main/java/cn/haitaoss/tinyspringmvc/framework/handlerMapping/Controller.java
@@ -1,0 +1,10 @@
+package cn.haitaoss.tinyspringmvc.framework.handlerMapping;
+
+/**
+ * @author haitao.chen
+ * email haitaoss@aliyun.com
+ * date 2021-04-25 11:22
+ * 所有Controller的实现类均会被ControllerHandlerMapping注册
+ */
+public interface Controller {
+}