+++ changelog.md
@@ -355,3 +355,69 @@ private void doDispatch(HttpServletRequest req, HttpServletResponse resp) throws
 }
 ```
 
+# step-3.0-solve-step2Problems
+
+> 解决第二步里面的问题
+
+- handlerRegistry 泛型应该是Object
+
+```java
+public abstract class AbstractHandlerMapping implements HandlerMapping {
+    protected static HashMap<String, Object> handlerRegistry = new HashMap<>();
+    protected static List<HandlerInterceptor> handlerInterceptors = new ArrayList<>();
+    protected ApplicationContext mvcContext;
+}
+```
+
+- HandlerExecutionChain 中的handler应该是Object 类型而不是RequestMappingHandler
+
+```java
+public class HandlerExecutionChain {
+    private Object handler;
+    private List<HandlerInterceptor> interceptors;
+
+    public Object getHandler() {
+        return handler;
+    }
+
+    public void setHandler(Object handler) {
+        this.handler = handler;
+    }
+
+    public List<HandlerInterceptor> getInterceptors() {
+        return interceptors;
+    }
+
+    public void setInterceptors(List<HandlerInterceptor> interceptors) {
+        this.interceptors = interceptors;
+    }
+}
+```
+
+- 调用的时候需要判断handler 的类型在调用
+
+```java
+private void doDispatch(HttpServletRequest req, HttpServletResponse resp) throws Exception {
+  // 找到请求对应的handler
+  HandlerExecutionChain handlerExecutionChain = handlerMapping.getHandler(req);
+  Object handler = handlerExecutionChain.getHandler();
+  List<HandlerInterceptor> handlerInterceptors = handlerExecutionChain.getInterceptors();
+
+  // 拦截器执行特点： 先执行全部的preHandler方法。只要执行的preHandler方法没有返回false 那么对应的afterCompletion一定会执行。
+  for (int i = 0; i < handlerInterceptors.size(); i++) {
+    HandlerInterceptor handlerInterceptor = handlerInterceptors.get(i);
+    if (!handlerInterceptor.preHandle(req, resp, handler)) {
+      for (int j = i - 1; j >= 0; j--) {
+        handlerInterceptors.get(j).afterCompletion(req, resp, handler, new Exception());
+      }
+      break;
+    }
+  }
+  // 只要某一个HandlerInterceptor 的 preHandler 返回了false，那么不应该调用目标方法（这里还没有实现这个功能）
+  // 至于如何传参，就是HandlerAdapter的事情了
+  if (handler instanceof RequestMappingHandler) {
+    ((RequestMappingHandler) handler).getMethod().invoke(((RequestMappingHandler) handler).getBean(), null);
+  }
+}
+```
+